Convert to a Reversible Microsoft Word Document

Format for converting from R Markdown to a Microsoft Word Document that can be reversed using undoc() after editing in Word.

rdocx_reversible(highlight_outputs = FALSE, wrap = 80,
  margins = NULL, ...)

Arguments

highlight_outputs
whether to highlight outputs from chunks and inline code in the final document

wrap
when round-tripping the document, at what width to wrap the markdown output? See undoc().

margins
page margin size. Can be a single value or a named vector with values, top, bottom, left, right, gutter, header, and footer. If NULL defaults to the reference document.

...
other parameters passed to rmarkdown::word_document()