R Markdown format for Reversible Reproducible Word Documents

Format for converting from R Markdown to a Microsoft Word Document that can be reversed using dedoc() after editing in Word.

redoc(highlight_outputs = TRUE, wrap = 80, margins = NULL,
  line_numbers = NULL, comment_author = NULL, keep_md = FALSE,
  wrappers = list(htmlcommentwrap, latexwrap, rawblockwrap, rawspanwrap,
  cmwrap, citationwrap), diagnostics = TRUE, ...)

Arguments

highlight_outputs
whether to highlight outputs from chunks and inline code in the final document

wrap
when round-tripping the document, at what width to wrap the markdown output? See dedoc().

margins
page margin size. Can be a single value or a named vector with values, top, bottom, left, right, gutter, header, and footer. If NULL defaults to the reference document.

line_numbers
either TRUE or list with any of the arguments start, by, restart, and distance

comment_author
The name to attribute any CriticMarkup tracked changes to. Defaults to whoami::fullname().

keep_md
whether to keep the intermediate markdown document

wrappers
a list of wrapper functions to capture text to protect when rendering and de-rendering. See make_wrapper().

diagnostics
Whether to embed diagnostic information in the output file. If TRUE, this will save session information and the current pandoc and (if used) RStudio versions inside the Word document for later bug-checking.

...
other parameters passed to rmarkdown::word_document()